Extract dbcmd.go into a new package under lib/client/db/dbcmd
